Assembly-report-automation/requirements/common.txt
SQLAlchemy==1.3.15
pytz==2020.1
pyOpenSSL==19.1.0
ndg-httpsclient==0.5.1
pyasn1==0.4.8
apache-airflow[crypto,celery,postgres,hive,jdbc,mysql,ssh]==1.10.9
redis==3.2
selenium==3.8.0
# boto3 is required for AWS service
boto3==1.14.20
# flask-bcrypt package is required for airflow.contrib.auth.backends.password_auth to authenticate the user on airflow
flask-bcrypt==0.7.1
# xlrd package is required for reading excel file
xlrd==1.2.0
# openpyxl package is required for writing into excel file
openpyxl==3.0.4
# package required for data validation
inflect==4.1.0
